+++ Runtime/GameEvents/Generic/BaseGameEventListener.cs
@@ -15,6 +15,29 @@ namespace GameEvents.Generic
         [Tooltip("Called when the listener is triggered with an argument")]
         private UnityEvent onGameEvent = default;
 
+        /// <summary>
+        ///     Get or set the underlying GameEvent.
+        /// </summary>
+        public TGameEvent GameEvent
+        {
+            get => gameEvent;
+            set => gameEvent = value;
+        }
+
+        /// <summary>
+        ///     Get or set the underlying UnityEvent.
+        /// </summary>
+        public UnityEvent OnGameEvent
+        {
+            get => onGameEvent;
+            set => onGameEvent = value;
+        }
+
+        public void RaiseGameEvent()
+        {
+            onGameEvent.Invoke();
+        }
+
         private void OnEnable()
         {
             if (gameEvent == null)
@@ -35,11 +58,6 @@ namespace GameEvents.Generic
             }
 
             gameEvent.UnregisterListener(this);
-        }
-
-        public void OnGameEvent()
-        {
-            onGameEvent.Invoke();
         }
     }
 }