+++ Runtime/GameEvents/Vector2/Vector2GameEvent.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 namespace GameEvents.Vector2
 {
     [CreateAssetMenu(fileName = "Vector2GameEvent", menuName = "Game Events/Vector2 Game Event")]
-    public class Vector3GameEvent : ArgumentGameEvent<UnityEngine.Vector2>
+    public class Vector2GameEvent : ArgumentGameEvent<UnityEngine.Vector2>
     {
     }
 }